swap default benchmarksURL with unpkg URL
