+++ package.json
@@ -15,8 +15,7 @@
     "build": "rollup -c rollup/config.lib.js",
     "example": "rollup -w -c rollup/config.dev.ts",
     "parse-analytics": "node ./scripts/analytics_parser.js",
-    "update-benchmarks": "rimraf benchmarks && mkdir -p benchmarks && ts-node -O '{\"module\":\"commonjs\"}' ./scripts/update_benchmarks.ts && npm run merge-benchmarks && tar -czvf benchmarks.tar.gz benchmarks/*.json",
-    "merge-benchmarks": "ts-node -O '{\"module\":\"commonjs\"}' ./scripts/merge_benchmarks.ts"
+    "update-benchmarks": "rimraf benchmarks && mkdir -p benchmarks && ts-node -O '{\"module\":\"commonjs\"}' ./scripts/update_benchmarks.ts && tar -czvf benchmarks.tar.gz benchmarks/*.json"
   },
   "author": "Tim van Scherpenzeel",
   "dependencies": {
